change np.int to int and np.float to float to remove warning from python3.9

+++ general/cubing.py
@@ -225,7 +225,7 @@ else:
 
 # Some global constants:
 
-epsilon = np.finfo(np.float).eps
+epsilon = np.finfo(float).eps
 # Store the value of epsilon for quick access.
 
 def get_object_names(infile):
